Github_weather/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scipy>=1.7.0
xarray>=0.19.0
cfgrib>=0.9.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
ecmwf-api-client>=1.6.0
requests>=2.25.0
